+++ solvers/ortoolsmp/main.cc
@@ -1,5 +1,5 @@
 #include "mp/backend-app.h"
-#include "ortoolsbackend.h"
+#include "ortoolsmpbackend.h"
 
 /// Declare a backend factory
 std::unique_ptr<mp::BasicBackend> CreateOrtoolsBackend();
