+++ solvers/ortoolsmp/ortoolsmp-lib.c
@@ -1,4 +1,4 @@
-#include "ortools/ortools-ampls-c-api.h"
+#include "ortoolsmp-ampls-c-api.h"
 
 #ifdef _WIN32
 #define APIEXPORT __declspec(dllexport)
