+++ solvers/ortoolsmp/ortoolsmpmodelapi.h
@@ -4,9 +4,9 @@
 #include <memory>
 
 #include "mp/env.h"
-#include "ortoolscommon.h"
+#include "ortoolsmpcommon.h"
 #include "mp/flat/model_api_base.h"
-#include "mp/flat/constraints_std.h"
+#include "mp/flat/constr_std.h"
 
 namespace mp {
